(#152) Padroniza responsividade da página Palavras-Chave
