+++ puma/src/components/palavras-chave/ListaKeyWords/ListaConsultaKeywords.vue
@@ -1,0 +1,5 @@
+<template src="./listaConsultaKeywords.html"></template>
+
+<script src="./listaConsultaKeywords.js"></script>
+
+<style src="./listaConsultaKeywords.css" scoped></style>
