+++ puma/src/components/palavras-chave/keyWords.js
@@ -1,8 +1,12 @@
-/* eslint-disable*/
 import KeywordService from '../../services/KeywordService';
 import SubjectService from '../../services/SubjectService';
+import ListaConsultaKeyWords from './ListaKeyWords/ListaConsultaKeywords.vue';
 
 export default {
+  components: {
+    ListaConsultaKeyWords,
+  },
+
   data() {
     return {
       kwNameAlreadyExist: false,
@@ -12,7 +16,6 @@ export default {
       openModalRegister: false,
       openModalEdit: false,
       currentKeyWord: {},
-      keywordDelete: null,
       selected: null,
       form: { keywordName: '', selectedSubject: null },
       idKeywordEdit: '',
@@ -26,34 +29,6 @@ export default {
       keyWords: [],
       keywordsInfo: {},
       tableKeywordSubject: [],
-      subjectsFields: [
-        {
-          key: 'id',
-          label: 'idDisciplina',
-        },
-        {
-          key: 'name',
-          label: 'subjectname',
-        },
-      ],
-      keyWordsFields: [
-        {
-          key: 'keywordid',
-          label: 'ITEM',
-        },
-        {
-          key: 'keyword',
-          label: 'PALAVRA-CHAVE',
-        },
-        {
-          key: 'subjectname',
-          label: 'DISCIPLINA',
-        },
-        {
-          key: 'actions',
-          label: '',
-        },
-      ],
     };
   },
   created() {
@@ -78,24 +53,6 @@ export default {
       }
     },
 
-    allowEdit(keyword) {
-      // return true;
-      const { isAdmin, userId } = this.$store.getters.user;
-      if (isAdmin) return true;
-
-      let flag = false;
-
-      Object.values(this.keywordsInfo[keyword.keywordid]).forEach((profId) => {
-        if (userId === profId) {
-          flag = true;
-        }
-      });
-      if (flag) {
-        return true;
-      }
-      return false;
-    },
-
     async getKeywords() {
       try {
         const { data } = await this.keywordService.getKeywords();
@@ -104,18 +61,25 @@ export default {
         Object.keys(data).forEach((key) => {
           this.keywordsInfo[data[key].keywordid] = data[key].array_agg;
         });
-      } catch (error) { }
+      } catch (error) {
+        this.makeToast('ERRO', 'Falha ao carregar as palavras-chave', 'danger');
+      }
     },
 
     async getSubjects() {
       try {
-        const allSubjects = (await this.subjectService.getSubjects()).data.map((s) => ({ value: s.subjectid, text: s.name }));
+        const allSubjects = (await this.subjectService.getSubjects()).data.map(
+          (s) => ({ value: s.subjectid, text: s.name }
+          ),
+        );
         this.subjectsForm = JSON.parse(JSON.stringify(allSubjects));
         this.subjectsForm.unshift({ value: null, text: 'Escolha a disciplina', disabled: true });
         this.subjects = JSON.parse(JSON.stringify(allSubjects));
         this.subjects.unshift({ value: 0, text: 'Todas as Disciplinas' });
         this.subjects.unshift({ value: null, text: 'Escolha a disciplina', disabled: true });
-      } catch (error) { }
+      } catch (error) {
+        this.makeToast('ERRO', 'Falha ao carregar as disciplinas', 'danger');
+      }
     },
 
     async getSubjectsProfessor() {
@@ -125,22 +89,33 @@ export default {
         Object.keys(this.keyWords).forEach((key) => {
           Object.values(this.keyWords[key].array_agg).forEach((allowId) => {
             if (allowId === userId) {
-              subjectByKeywords.push({ value: this.keyWords[key].subjectid, text: this.keyWords[key].subjectname });
+              subjectByKeywords.push({
+                value: this.keyWords[key].subjectid,
+                text: this.keyWords[key].subjectname,
+              });
             }
           });
         });
-        this.subjectsForm = [...new Set(subjectByKeywords.map((o) => JSON.stringify(o)))].map((s) => (JSON.parse(s)));
+        this.subjectsForm = [...new Set(subjectByKeywords.map((o) => JSON.stringify(o)))].map(
+          (s) => (JSON.parse(s)),
+        );
         this.subjectsForm.unshift({ value: null, text: 'Escolha a disciplina', disabled: true });
-      } catch (erro) { }
+      } catch (erro) {
+        this.makeToast('ERRO', 'Falha ao carregar as disciplinas', 'danger');
+      }
     },
 
     async getFilterProfessor() {
       try {
-        const allSubjects = (await this.subjectService.getSubjects()).data.map((s) => ({ value: s.subjectid, text: s.name }));
+        const allSubjects = (await this.subjectService.getSubjects()).data.map(
+          (s) => ({ value: s.subjectid, text: s.name }),
+        );
         this.subjects = JSON.parse(JSON.stringify(allSubjects));
         this.subjects.unshift({ value: 0, text: 'Todas as Disciplinas' });
         this.subjects.unshift({ value: null, text: 'Escolha a disciplina', disabled: true });
-      } catch (error) { }
+      } catch (error) {
+        this.makeToast('ERRO', 'Falha ao carregar Professor', 'danger');
+      }
     },
 
     filter(subjectId) {
@@ -148,40 +123,23 @@ export default {
       if (subjectId === 0) {
         this.tableKeywordSubject = this.keyWords;
       } else {
-        this.tableKeywordSubject = this.tableKeywordSubject.filter((d) => d.subjectid === subjectId);
+        this.tableKeywordSubject = this.tableKeywordSubject.filter(
+          (d) => d.subjectid === subjectId,
+        );
       }
     },
 
     addKeyword() {
       this.openModalRegister = true;
+      this.getKeywords();
       this.form = { keywordName: '', selectedSubject: null };
     },
 
-    editKeyword(keyword) {
-      this.openModalEdit = true;
-      this.idKeywordEdit = keyword.keywordid;
-      this.form = { keywordName: keyword.keyword, selectedSubject: keyword.subjectid };
-    },
-
-    deleteKeyword(keyWord) {
-      this.$store.commit('OPEN_CONFIRM_MODAL', {
-        title: 'EXCLUIR PALAVRA-CHAVE',
-        content: 'Confirmar exclusão da palavra-chave ?',
-        cancelButton: {
-          text: 'Cancelar', variant: 'outline-danger',
-          onClick: () => this.$store.commit('CLOSE_CONFIRM_MODAL'),
-        },
-        okButton: {
-          text: 'Confirmar', variant: 'danger',
-          onClick: () => { this.handleDelete(); this.$store.commit('CLOSE_CONFIRM_MODAL'); },
-        },
-      });
-      this.keywordDelete = keyWord.keywordid;
-    },
-
     keywordNameAlreadyExist() {
-      const currentKeyword = this.form.keywordName;
-      this.kwNameAlreadyExist = this.tableKeywordSubject.some((k) => this.treatKeyword(k.keyword) === this.treatKeyword(this.form.keywordName),);
+      // const currentKeyword = this.form.keywordName;
+      this.kwNameAlreadyExist = this.tableKeywordSubject.some(
+        (k) => this.treatKeyword(k.keyword) === this.treatKeyword(this.form.keywordName),
+      );
     },
 
     treatKeyword(keyword) { return keyword.split(' ').join('').toLowerCase(); },
@@ -211,41 +169,16 @@ export default {
       }
     },
 
-    async handleEdit() {
-      try {
-        const isFormValid = await this.$refs.observer.validate();
-        if (isFormValid && this.kwNameAlreadyExist === false) {
-          this.$store.commit('OPEN_LOADING_MODAL', { title: 'Enviando...' });
-          const response = await this.keywordService.updateKeyword(this.idKeywordEdit, this.form.keywordName);
-          const idKeywordUpdated = response.data.keywordid;
-          await this.keywordService.updateSubjectKeyword(idKeywordUpdated, this.form.selectedSubject);
-          this.openModalEdit = false;
-          this.$store.commit('CLOSE_LOADING_MODAL');
-          this.makeToast('SUCESSO', 'Palavra-Chave editada com sucesso!', 'success');
-          this.getKeywords();
-        }
-      } catch (error) {
-        this.openModalEdit = false;
-        this.$store.commit('CLOSE_LOADING_MODAL');
-        this.makeToast('ERRO', 'Infelizmente houve um erro ao tentar editar a palavra-chave', 'danger');
-      }
-    },
-
-    async handleDelete() {
-      try {
-        this.$store.commit('OPEN_LOADING_MODAL', { title: 'Enviando...' });
-        await this.keywordService.deleteKeyword(this.keywordDelete);
-        await this.getKeywords();
-        this.$store.commit('CLOSE_LOADING_MODAL');
-        this.makeToast('SUCESSO', 'Palavra-chave excluída com sucesso!', 'success');
-      } catch (error) {
-        this.$store.commit('CLOSE_LOADING_MODAL');
-        this.makeToast('ERRO', 'Infelizmente houve um erro ao tentar excluir a palavra-chave', 'danger');
-      }
-    },
-
-    makeToast(title, message, variant) {
-      this.$bvToast.toast(message, { title: title, variant: variant, solid: true, autoHideDelay: 4000, });
+    makeToast(thisTitle, message, thisVariant) {
+      this.$bvToast.toast(
+        message,
+        {
+          title: thisTitle,
+          variant: thisVariant,
+          solid: true,
+          autoHideDelay: 4000,
+        },
+      );
     },
   },
 };
